Enable SSR for home client
